+++ benches/array_blosc.rs
@@ -1,7 +1,7 @@
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, Throughput};
 use zarrs::{
     array::codec::BloscCodec,
-    metadata::v3::codec::blosc::{BloscCompressionLevel, BloscCompressor, BloscShuffleMode},
+    metadata::v3::array::codec::blosc::{BloscCompressionLevel, BloscCompressor, BloscShuffleMode},
 };
 
 fn array_blosc_write_all(c: &mut Criterion) {