+++ src/array/array_async_writable.rs
@@ -6,10 +6,8 @@ use crate::{
     array::ArrayBytes,
     array_subset::ArraySubset,
     config::{global_config, MetadataEraseVersion},
-    storage::{
-        meta_key_v2_array, meta_key_v2_attributes, meta_key_v3, AsyncBytes,
-        AsyncWritableStorageTraits, StorageError, StorageHandle,
-    },
+    node::{meta_key_v2_array, meta_key_v2_attributes, meta_key_v3},
+    storage::{AsyncBytes, AsyncWritableStorageTraits, StorageError, StorageHandle},
 };
 
 use super::{