+++ src/array/array_sync_writable.rs
@@ -7,10 +7,8 @@ use crate::{
     array::ArrayBytes,
     array_subset::ArraySubset,
     config::{global_config, MetadataEraseVersion},
-    storage::{
-        meta_key_v2_array, meta_key_v2_attributes, meta_key_v3, Bytes, StorageError, StorageHandle,
-        WritableStorageTraits,
-    },
+    node::{meta_key_v2_array, meta_key_v2_attributes, meta_key_v3},
+    storage::{Bytes, StorageError, StorageHandle, WritableStorageTraits},
 };
 
 use super::{