+++ src/array/chunk_key_encoding/v2.rs
@@ -102,7 +102,7 @@ impl ChunkKeyEncodingTraits for V2ChunkKeyEncoding {
 
 #[cfg(test)]
 mod tests {
-    use crate::{node::NodePath, storage::data_key};
+    use crate::node::{data_key, NodePath};
 
     use super::*;