+++ src/node/node_sync.rs
@@ -5,13 +5,15 @@ use crate::{
     group::GroupMetadata,
     metadata::{v2::GroupMetadataV2, v3::NodeMetadataV3},
     storage::{
-        discover_children, meta_key_v2_array, meta_key_v2_attributes, meta_key_v2_group,
-        meta_key_v3, ListableStorageTraits, ReadableStorageTraits, StorageError, StoreKey,
+        discover_children, ListableStorageTraits, ReadableStorageTraits, StorageError, StoreKey,
         StorePrefix,
     },
 };
 
-use super::{Node, NodeMetadata, NodePath};
+use super::{
+    meta_key_v2_array, meta_key_v2_attributes, meta_key_v2_group, meta_key_v3, Node, NodeMetadata,
+    NodePath,
+};
 
 fn get_metadata_v3<TStorage: ?Sized + ReadableStorageTraits + ListableStorageTraits>(
     storage: &Arc<TStorage>,